Added Validation Exception

+++ Assignment1/src/service/employee.service.ts
@@ -1,5 +1,8 @@
 import EmployeeRepository from "../repository/employee.repository";
 import Employee from "../entity/employee.entity";
+import Address from "../entity/address.entity";
+import { test } from "node:test";
+import HttpException from "../exceptions/http.exception";
 
 
 class EmployeeService{
@@ -10,19 +13,33 @@ class EmployeeService{
         return this.employeeRepository.findAllEmployee();
     }
 
-    getEmployeeByID(id: number): Promise<Employee>{
-        return this.employeeRepository.findEmployeeByID(id);
+    async getEmployeeByID(id: number): Promise<Employee>{
+        const emp = await this.employeeRepository.findEmployeeByID(id);
+        if(!emp){
+            throw new HttpException(404, `Employee not found with ID ${id}`);
+        }
+        return emp;
+
     }
 
-    addNewEmployee(name: string, email: string): Promise<Employee>{
+    addNewEmployee(name: string, email: string, address:any): Promise<Employee>{
         const newEmp = new Employee();
         newEmp.name = name;
         newEmp.email = email;
+
+        const newAddress = new Address();
+        newAddress.line1 = address.line1;
+        newAddress.pincode = address.pincode;
+        newEmp.address = newAddress;
+
         return this.employeeRepository.saveNewEmp(newEmp);
     }
 
     async updateEmployee(id: number, name: string, email: string): Promise<Employee>{
         const emp = await this.getEmployeeByID(id);
+        if(!emp){
+            throw new HttpException(404, `Employee not found with ID ${id}`);
+        }
         emp.name = name;
         emp.email = email;
         return this.employeeRepository.updateEmp(emp);
@@ -30,6 +47,9 @@ class EmployeeService{
 
     async deleteEmployeeByID(id: number): Promise<Employee>{
         const emp = await this.getEmployeeByID(id = id);
+        if(!emp){
+            throw new HttpException(404, `Employee not found with ID ${id}`);
+        }
         return this.employeeRepository.deleteEmp(emp);
     }
 }
